Move metadata selector test
